Port more code

+++ Engine/CatFile.cs
@@ -99,7 +99,7 @@ internal class CatFile : BinaryReader
             }
             else
             {
-                _size[i] = (uint)(_size[i] + namesize + 1);
+                _size[i] = _size[i] + namesize + 1;
             }
         }
